fix: remove idx types, move test utils
