+++ apps/integration/package.json
@@ -28,7 +28,6 @@
     "@babel/plugin-transform-runtime": "^7.12.13",
     "@babel/preset-env": "^7.12.13",
     "@babel/preset-typescript": "^7.12.13",
-    "@ceramicstudio/idx-tools": "^0.11.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-loader": "^8.2.2",
     "css-loader": "^5.2.6",
