+++ packages/manager/package.json
@@ -43,10 +43,6 @@
     "@ceramicnetwork/blockchain-utils-linking": "^1.2.0",
     "@ceramicnetwork/http-client": "^1.4.2",
     "@ceramicnetwork/stream-caip10-link": "^1.1.5",
-    "@datamodels/3id-keychain": "^0.1.2",
-    "@datamodels/identity-accounts-crypto": "^0.1.2",
-    "@datamodels/identity-accounts-web": "^0.1.2",
-    "@datamodels/identity-profile-basic": "^0.1.2",
     "@ethersproject/hdnode": "^5.0.8",
     "@glazed/devtools": "^0.1.3",
     "@glazed/did-datastore": "^0.1.4",
@@ -62,6 +58,10 @@
   "devDependencies": {
     "@3id/test-utils": "^0.2.5",
     "@ceramicnetwork/common": "^1.7.0",
+    "@datamodels/3id-keychain": "^0.1.2",
+    "@datamodels/identity-accounts-crypto": "^0.1.2",
+    "@datamodels/identity-accounts-web": "^0.1.2",
+    "@datamodels/identity-profile-basic": "^0.1.2",
     "did-jwt": "^5.1.2",
     "ipfs-repo": "^9.1.1",
     "prom-client": "^13.1.0",
